requirements.txt
certifi==2021.10.8
cycler==0.11.0
kiwisolver==1.3.2
matplotlib==3.4.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095651905/work
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
opencv-python==4.5.4.58
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
six @ file:///tmp/build/80754af9/six_1623709665295/work
torch==1.10.0
torchaudio==0.10.0
torchfile==0.1.0
torchvision==0.11.1
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
